move get_model to utils to not use methods from train.py in evaluation scripts
